requirements.txt
anyio==3.6.2
click==8.1.3
colorama==0.4.6
fastapi==0.89.1
h11==0.14.0
httptools==0.5.0
idna==3.4
pip==21.3.1
pydantic==1.10.4
python-dotenv==0.21.1
PyYAML==6.0
setuptools==60.2.0
sniffio==1.3.0
starlette==0.22.0
typing-extensions==4.4.0
uvicorn==0.20.0
watchfiles==0.18.1
websockets==10.4
wheel==0.37.1
sqlalchemy
mysql-connector-python
python-multipart
python-jose[cryptography]